requirements.txt
# Pursuit-Evasion Hybrid RL Dependencies
# =======================================

# Core RL
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Numerical computing
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0

# Configuration
pyyaml>=6.0

# Progress bars
rich>=13.0.0
tqdm>=4.65.0

# Utilities
tensorboard>=2.13.0
